websocket/python/requirements.txt
pyserial==3.4
websockets==8.1